refactored user service using functions - refactored protector function into user service instead of (auth)
