+++ app.js
@@ -8,7 +8,7 @@ const postRouter = require('./posts/post.routes');
 const imageRouter = require('./images/image.routes');
 const userRouter = require('./users/user.routes');
 const feedbackRouter = require('./feedbacks/feedback.routes');
-const { protector } = require('./auth/auth.controller');
+const { protector } = require('./users/user.controller');
 const errorHandler = require('./middleware/errorhandler');
 const swaggerUi = require('swagger-ui-express');
 const swaggerDocument = require('./m3ntorship-Posts-1.0.0-swagger.json');
@@ -44,7 +44,7 @@ app.get('/protected', (req, res) => {
 
 app.use('/posts', postRouter);
 app.use('/images', imageRouter);
-app.use('/user', userRouter);
+app.use('/users', userRouter);
 app.use('/feedbacks', feedbackRouter);
 
 app.use(
