--- auth/auth.controller.js
@@ -1,20 +1,0 @@
-const appError = require('../util/appError');
-const admin = require('firebase-admin');
-const { resolve } = require('path');
-const User = require('../users/user.model');
-const catchAsync = require('../util/catchAsync');
-const { protector } = require('@m3ntorship/pickly-protector');
-const serviceAccount = resolve('secrets', 'service-account.json');
-
-const userEnricher = async user => {
-  let mongoUser = await User.findOne({ email: user.tokeninfo.email });
-  if (!mongoUser)
-    mongoUser = await User.create({
-      name: user.tokeninfo.name,
-      email: user.tokeninfo.email,
-      userImage: user.tokeninfo.picture
-    });
-  return mongoUser;
-};
-
-module.exports.protector = catchAsync(protector(serviceAccount, userEnricher));
