+++ users/user.service.js
@@ -1,59 +1,58 @@
-const Post = require('../posts/post.model');
-const AppError = require('../util/appError');
-const catchAsync = require('../util/catchAsync');
+const { resolve } = require('path');
+const { protector } = require('@m3ntorship/pickly-protector');
 const User = require('./user.model');
+const Post = require('../posts/post.model');
+const catchAsync = require('../util/catchAsync');
+
+const serviceAccount = resolve('secrets', 'service-account.json');
+
+const userEnricher = async user => {
+  let mongoUser = await User.findOne({ email: user.tokeninfo.email });
+  if (!mongoUser)
+    mongoUser = await User.create({
+      name: user.tokeninfo.name,
+      email: user.tokeninfo.email,
+      userImage: user.tokeninfo.picture
+    });
+  return mongoUser;
+};
+
+const getPosts = opts => {
+  const userPosts = Post.find(opts).populate({
+    path: 'resources',
+    model: 'resources',
+    populate: {
+      path: 'images',
+      model: 'image',
+      select: 'name url',
+      populate: {
+        path: 'votes',
+        model: 'Votes',
+        select: 'count  updatedAt'
+      }
+    }
+  });
+  return userPosts;
+};
 
 exports.userService = {
-  getPosts() {
+  protector() {
+    return catchAsync(protector(serviceAccount, userEnricher));
+  },
+  getUserPosts() {
     return catchAsync(async (req, res, next) => {
-      // owner
+      let userPosts;
       if (req.user.mongouser._id.toString() === req.params.userId) {
-        // return all Posts
-        const userPosts = await Post.find({
+        userPosts = await getPosts({
           author: req.params.userId
-        }).populate({
-          path: 'resources',
-          model: 'resources',
-          populate: {
-            path: 'images',
-            model: 'image',
-            select: 'name url',
-            populate: {
-              path: 'votes',
-              model: 'Votes',
-              select: 'count  updatedAt'
-            }
-          }
         });
-        // const userPosts = await User.findById(req.params.userId)
-        //   .select('name email userImage posts')
-        //   .populate({
-        //     path: 'posts',
-        //     model: 'Post',
-        //     populate: {
-        //       path: 'resources',
-        //       model: 'resources',
-        //       populate: {
-        //         path: 'images',
-        //         model: 'image',
-        //         select: 'name url',
-        //         populate: {
-        //           path: 'votes',
-        //           model: 'Votes',
-        //           select: 'count  updatedAt'
-        //         }
-        //       }
-        //     }
-        //   });
-        return res.json(userPosts);
       } else {
-        return next(
-          new AppError(
-            `You are not allowed to view this profile posts at the current moment`,
-            403
-          )
-        );
+        userPosts = await getPosts({
+          author: req.params.userId,
+          isAnonymous: false
+        });
       }
+      return res.json(userPosts);
     });
   }
 };
